Restructure plugin to a proper package structure
