+++ pyproject.toml
@@ -24,4 +24,3 @@ lint = ["black"]
 
 [tool.setuptools.packages.find]
 include = ["llm_ollama*"]
-exclude = ["cache"]
